# contract/openapi.json
{
  "components": {
    "schemas": {
      "AssignPresetRequest": {
        "properties": {
          "preset_id": {
            "description": "Null clears the assignment.",
            "format": "uuid",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "type": "object"
      },
      "AssignPresetResponse": {
        "properties": {
          "assigned_at": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "assigned_preset_id": {
            "format": "uuid",
            "type": [
              "string",
              "null"
            ]
          },
          "device_id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "device_id"
        ],
        "type": "object"
      },
      "AssignedPreset": {
        "description": "The server-chosen default preset, named well enough to select locally.",
        "properties": {
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "version": {
            "format": "int32",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "version"
        ],
        "type": "object"
      },
...
      "HeartbeatRequest": {
        "properties": {
          "active_preset_name": {
            "description": "The preset the device currently runs under, acknowledging an assignment.",
            "type": [
              "string",
              "null"
            ]
          },
          "active_preset_version": {
            "format": "int32",
            "type": [
              "integer",
              "null"
            ]
          },
          "gui_version": {
            "type": [
              "string",
              "null"
            ]
          },
          "library_version": {
            "type": [
              "string",
              "null"
            ]
          },
          "platform": {
            "type": [
              "string",
              "null"
            ]
          },
          "preset_schema_version": {
            "description": "Which `preset-schema.json` revision this installation understands.",
            "format": "int32",
            "type": [
              "integer",
              "null"
            ]
          },
          "system_profile": {
            "description": "Arbitrary JSON object describing the hardware, stored as sent."
          }
        },
        "type": "object"
      },
      "HeartbeatResponse": {
        "properties": {
          "assigned_preset": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AssignedPreset",
                "description": "Null when nothing is assigned, or the assigned preset has been deleted."
              }
            ]
          }
        },
        "type": "object"
      },
...
      "MintConnectCodeRequest": {
        "properties": {
          "device_name": {
            "description": "The name the redeeming device takes, so a device's name has one origin: the\nperson minting the code names the installation they are about to enrol.",
            "type": "string"
          }
        },
        "required": [
          "device_name"
        ],
        "type": "object"
      },
      "MintConnectCodeResponse": {
        "properties": {
          "code": {
            "description": "The string to paste into the desktop application. Shown once.",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "code",
          "expires_at"
        ],
        "type": "object"
      },
...
      "RenameDeviceRequest": {
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
      "RenameDeviceResponse": {
        "properties": {
          "device_id": {
            "format": "uuid",
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "device_id",
          "name"
        ],
        "type": "object"
      },
      "RevokeResponse": {
        "properties": {
          "device_id": {
            "format": "uuid",
            "type": "string"
          },
          "revoked_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "device_id",
          "revoked_at"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "device_token": {
        "description": "Device token issued at enrolment, used by the desktop application",
        "scheme": "bearer",
        "type": "http"
      },
      "keycloak_jwt": {
        "bearerFormat": "JWT",
        "description": "Keycloak JWT, used by the web interface",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "ECEO Team"
    },
    "description": "Reef survey metadata registry",
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "title": "DeepReefMap API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/devices/connect-codes": {
      "post": {
        "description": "Interactive login only, so a device cannot invite further devices.",
        "operationId": "mint_code",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MintConnectCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MintConnectCodeResponse"
                }
              }
            },
            "description": "Code minted; shown to the operator once"
          },
          "400": {
            "description": "Empty device name"
          },
          "403": {
            "description": "Requires an interactive login"
          },
          "500": {
            "description": "PUBLIC_BASE_URL is not configured"
          }
        },
        "summary": "Mint a connect code for a desktop installation.",
        "tags": [
          "devices"
        ]
      }
    },
    "/devices/{device_id}/assign-preset": {
      "post": {
        "description": "The assignment travels in the next heartbeat response, and the device reports the\npreset it actually runs under in the request after that, so `active_preset_*` on\nthe device row says whether the assignment was acknowledged.",
        "operationId": "assign_preset",
        "parameters": [
          {
            "description": "Device to assign a preset to",
            "in": "path",
            "name": "device_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignPresetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignPresetResponse"
                }
              }
            },
            "description": "Assignment stored"
          },
          "403": {
            "description": "Not your device, or a device token"
          },
          "404": {
            "description": "No such device, or no such preset"
          }
        },
        "summary": "Choose a device's default preset. Members may assign to their own, administrators\nto anyone's.",
        "tags": [
          "devices"
        ]
      }
    },
    "/devices/{device_id}/rename": {
      "post": {
        "description": "Interactive login only. A device that could rename itself would make its own\nattribution editable.",
        "operationId": "rename",
        "parameters": [
          {
            "description": "Device to rename",
            "in": "path",
            "name": "device_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameDeviceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RenameDeviceResponse"
                }
              }
            },
            "description": "Device renamed"
          },
          "400": {
            "description": "Empty name"
          },
          "403": {
            "description": "Not your device, or a device token"
          },
          "404": {
            "description": "No such device"
          }
        },
        "summary": "Rename a device. Members may rename their own, administrators anyone's.",
        "tags": [
          "devices"
        ]
      }
    },
    "/devices/{device_id}/revoke": {
      "post": {
        "operationId": "revoke",
        "parameters": [
          {
            "description": "Device to revoke",
            "in": "path",
            "name": "device_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevokeResponse"
                }
              }
            },
            "description": "Device revoked"
          },
          "403": {
            "description": "Not your device"
          },
          "404": {
            "description": "No such device"
          }
        },
        "summary": "Revoke a device. Members may revoke their own, administrators anyone's.",
        "tags": [
          "devices"
        ]
      }
    },
...
    "/sync/heartbeat": {
      "post": {
        "description": "Identity comes from the credential and never from the body, so a device cannot\nreport on a sibling's behalf.",
        "operationId": "heartbeat",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeartbeatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HeartbeatResponse"
                }
              }
            },
            "description": "Report stored; the response names the assigned preset"
          },
          "400": {
            "description": "system_profile is not a JSON object"
          },
          "403": {
            "description": "Requires a device token"
          }
        },
        "summary": "Update the calling device's own record and learn its assigned preset.",
        "tags": [
          "devices"
        ]
      }
    },
...
  },
...
}